refactor: add validation tests
