+++ packages/crypto/__tests__/validation/rules/models/transactions/multi-signature.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const rule = require('../../../../lib/rules/models/transactions/multi-signature')
-const { constants, crypto, transactionBuilder } = require('@arkecosystem/crypto')
+const rule = require('../../../../../lib/validation/rules/models/transactions/multi-signature')
+const { crypto, constants, transactionBuilder } = require('../../../../../lib')
 const passphrase = 'passphrase 1'
 const publicKey = '+03e8021105a6c202097e97e6c6d650942d913099bf6c9f14a6815df1023dde3b87'
 const passphrases = [
