+++ packages/crypto/__tests__/validation/rules/models/transactions/transfer.test.js
@@ -1,7 +1,7 @@
 'use strict'
 
-const rule = require('../../../../lib/rules/models/transactions/transfer')
-const { constants, transactionBuilder } = require('@arkecosystem/crypto')
+const rule = require('../../../../../lib/validation/rules/models/transactions/transfer')
+const { constants, transactionBuilder } = require('../../../../../lib')
 const address = 'APnDzjtDb1FthuqcLMeL5XMWb1uD1KeMGi'
 const fee = 1 * constants.ARKTOSHI
 const amount = 10 * constants.ARKTOSHI
